requirements.txt
# Core Flask
flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3

# Flask-SocketIO and dependencies
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
bidict==0.22.1

# Environment variables
python-dotenv==1.0.0

# Google Generative AI
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.27.0
googleapis-common-protos==1.62.0

# Server
gunicorn==21.2.0
eventlet==0.35.2

# Additional dependencies
six==1.16.0
dnspython==2.5.0
greenlet==3.0.3